Event, Part, Main updates, Folder structure
Event and Part Class: added constructors with an id as a parameter. When an ID is defined, it is treated as the last value and the counter will start from the next ID.
Main: added an ArrayList for the Part objects. It is not used in this case, but might be useful in the future, if we need to generate a report from the loaded data. When reading from a file, the last Event's ID is used for the idCounter.
-Folder Structure: moved the lib folder to the root of the project. The table creation script has been moved to the scripts folder.
-README.md: updates to the time taken for each task. Description on where to find the database creation script.